Moved DOTCMS_CLIENT_TOKEN to angular example
