+++ core-web/libs/sdk/angular/src/index.ts
@@ -1,5 +1,3 @@
 export * from './lib/layout/dotcms-layout/dotcms-layout.component';
-export * from './lib/client-token/dotcms-client-token';
 export * from './lib/services/dotcms-context/page-context.service';
-export * from './lib/resolver/dotcms-page.resolver';
 export * from './lib/models';
