+++ examples/angular/src/app/app.routes.ts
@@ -1,6 +1,5 @@
 import { Routes } from '@angular/router';
 import { DotCMSPagesComponent } from './pages/pages.component';
-// import { DotCMSPageResolver } from '@dotcms/angular';
 import { DotCMSPageResolver } from './resolver/dotcms-page.resolver';
 
 export const routes: Routes = [
