+++ examples/angular/src/app/client-token/dotcms-client.ts
