+++ examples/angular/src/app/resolver/dotcms-page.resolver.ts
@@ -1,61 +1,55 @@
-import { Inject, Injectable, InjectionToken, inject } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
-import {
-  DOTCMS_CLIENT_TOKEN,
-  DotCMSPageAsset,
-  DotcmsNavigationItem,
-  PageContextService,
-} from '@dotcms/angular';
+import { ActivatedRouteSnapshot } from '@angular/router';
+import { DotCMSPageAsset, DotcmsNavigationItem } from '@dotcms/angular';
+import { inject } from '@angular/core';
+import { DOTCMS_CLIENT_TOKEN } from '../client-token/dotcms-client';
 
-interface DotCMSPageResolverResponse {
+/**
+ * This resolver is used to fetch the page and navigation data from dotCMS.
+ *
+ * @param {ActivatedRouteSnapshot} route
+ * @param {RouterStateSnapshot} _state
+ * @return {*}
+ */
+export const DotCMSPageResolver = async (
+  route: ActivatedRouteSnapshot,
+): Promise<{
   pageAsset: DotCMSPageAsset;
   nav: DotcmsNavigationItem;
-}
+}> => {
+  const client = inject(DOTCMS_CLIENT_TOKEN);
 
-@Injectable({
-  providedIn: 'root',
-})
-export class DotCMSPageResolver implements Resolve<DotCMSPageResolverResponse> {
-  // constructor(@Inject(DOTCMS_CLIENT_TOKEN) private client: DotCmsClient) {}
-  // some = new InjectionToken('some');
-  // someToken = inject(this.some) as any;
-  client = inject(DOTCMS_CLIENT_TOKEN as any) as any;
-  pageService = inject(PageContextService);
+  const url = route.url.map((segment) => segment.path).join('/');
+  const queryParams = route.queryParams;
 
-  async resolve(route: ActivatedRouteSnapshot) {
-    const url = route.url.map((segment) => segment.path).join('/');
-    const queryParams = route.queryParams;
+  const pageProps = {
+    path: url || 'index',
+    language_id: queryParams['language_id'],
+    mode: queryParams['mode'],
+    variantName: queryParams['variantName'],
+    'com.dotmarketing.persona.id':
+      queryParams['com.dotmarketing.persona.id'] || '',
+  };
 
-    const pageProps = {
-      path: url || 'index',
-      language_id: queryParams['language_id'],
-      mode: queryParams['mode'],
-      variantName: queryParams['variantName'],
-      'com.dotmarketing.persona.id':
-        queryParams['com.dotmarketing.persona.id'] || '',
-    };
+  const navProps = {
+    path: '/',
+    depth: 2,
+    languageId: queryParams['language_id'],
+  };
 
-    const navProps = {
-      path: '/',
-      depth: 2,
-      languageId: queryParams['language_id'],
-    };
+  const pageRequest = client.page.get(pageProps) as Promise<{
+    entity: DotCMSPageAsset;
+  }>;
+  const navRequest = client.nav.get(navProps) as Promise<{
+    entity: DotcmsNavigationItem;
+  }>;
 
-    const pageRequest = this.client.page.get(pageProps) as Promise<{
-      entity: DotCMSPageAsset;
-    }>;
-    const navRequest = this.client.nav.get(navProps) as Promise<{
-      entity: DotcmsNavigationItem;
-    }>;
+  const [pageResponse, navResponse] = await Promise.all([
+    pageRequest,
+    navRequest,
+  ]);
 
-    const [pageResponse, navResponse] = await Promise.all([
-      pageRequest,
-      navRequest,
-    ]);
+  const pageAsset = pageResponse.entity;
+  const nav = navResponse.entity;
 
-    const pageAsset = pageResponse.entity;
-    const nav = navResponse.entity;
-
-    return { pageAsset, nav };
-  }
-}
+  return { pageAsset, nav };
+};
